requirements.txt
Flask
requests
beautifulsoup4
bs4 
html5lib
lxml
selenium
webdriver_manager
ipykernel
pandas
sentence-transformers
faiss-cpu
langchain 
openai 
streamlit
langchain-community
pydantic
ipywidgets